Allow using more lifted functions when deriving Eq
Previously it wasn't possible to derive an `Eq1` instance for the
following type:

```
data Test a = Test (a, ())
```

The reason for this was that it requires using `liftEq2` in the
generated `Eq1` instance, and the library was previously restricted to
not use "more" lifted functions than the one it was generating.

This commit lifts that restriction.

It also adds some tests for this kind of situation.