Avoid nested ForallTs in `deriveVia`
This is unfortunately necessary due to GHC 9.2 becoming stricter about checking
for nested `forall`s in instance heads. Since we were relying on nested
`ForallT`s to make `deriveVia` generate instance heads with `forall`s, this
prevents us from generating instances with floating type variables. Unless
Template Haskell gains the ability to mention top-level `forall`s in instance
declarations some day, this is the best that we can do.

Fixes #34, and partially reverts #28 (the fix for #27) in the process.